requirements.txt
pandas~=3.0.0
polars
matplotlib
seaborn
numpy
scikit-learn
scipy
tqdm
cupy-cuda13x